renamed configuration.ts, lint elixirMain.ts
